【问题/需求】迁移CRNN样例提交文件 【原因分析】将本地modelzoo-dev中的CRNN样例改动同步到提交仓库 【修改描述】同步CRNN README、构建配置、测试数据、后处理与辅助脚本 【注意事项】不涉及 【涉及芯片】Hi3403V100 【自测用例】执行modelzoo-submit dry-run预览指定7个文件迁移范围 【自测结果】dry-run确认指定7个文件可同步到提交仓库

+++ samples/common/infer/post_process/crnn_postprocess.cpp
@@ -61,7 +61,7 @@ static Result GetAndSaveOutputWithBin(Infer::TensorBuf& outBuf, Infer::TensorDes
         std::vector<float> tempBuf(outData, outData + outSize);
         for (size_t i = 0; i < loopNum; ++i) {
             size_t offset = i * strideElemNum;
-            for (size_t index = 0; index < lastDim; index++) {
+            for (size_t index = 0; index < static_cast<size_t>(lastDim); index++) {
                 noStrideBuf.push_back(tempBuf[offset + index]);
             }
         }
@@ -125,7 +125,7 @@ void CalResult(const std::vector<float>& output, const std::string& fileName)
     }
     std::vector<int> charList;
     string cxt;
-    for (int i = 0; i < charIndeces.size(); ++i) {
+    for (size_t i = 0; i < charIndeces.size(); ++i) {
         if (charIndeces[i] != 0 && !(i > 0 && charIndeces[i - 1] == charIndeces[i])) {
             cxt += to_string(charIndeces[i]) + " ";
             charList.emplace_back(charIndeces[i]);
